Major clenup of the initial commit
Most of the previous files have been rearranged in a different
structure to match the new project scheme. Furthermore, some
of the previous capabilities have been incorporated inside other
modules, which will be addedd as extra dependencies.

At the moment, it is possibile interact with the PMC configuration
and read/write the system PMCs' state. No sample can be read.