Introduced PEBS support
PEBS logic has been introduced, so that it is possible to define
the reset value for each PMC that use PEBS and feed the hw logic
with the required policy. The PEBS init is done at module load and
the PMC configuration is load at setup time. The DS area patching
function create the old valu backup which is restored at module
unload. The module employes a finite number of buffers (per CPU)
so that PEBS will record for a finite number of time before using
up the entire space.